process_daily_narrative/requirements.txt
anthropic==0.34.1
feedparser==6.0.11
Flask==3.0.3
Flask_Cors==4.0.0
fpdf==1.7.2
google_api_python_client==2.129.0
html2text==2024.2.26
openai==1.42.0
Pillow==10.4.0
protobuf==5.27.3
pydub==0.25.1
PyGithub==2.3.0
PyPDF2==3.0.1
reportlab==4.2.0
Requests==2.32.3
google-cloud-storage
python-dateutil
firebase-admin
praw
beautifulsoup4
pytz
replicate